refactor: align height and inscription number types to u32
Match upstream ordinals/ord type conventions: use u32 for block
heights and inscription numbers throughout. Cast to u64 only at
RPC boundaries.

Closes #27

+++ src/chain.rs
@@ -57,7 +57,7 @@ impl Chain {
     }
   }
 
-  pub(crate) fn first_inscription_height(self) -> u64 {
+  pub(crate) fn first_inscription_height(self) -> u32 {
     match self {
       Self::Mainnet => 186920,
       Self::Regtest => 0,
