+++ src/options.rs
@@ -28,9 +28,9 @@ pub struct Options {
     long,
     help = "Don't look for inscriptions below <FIRST_INSCRIPTION_HEIGHT>."
   )]
-  pub(crate) first_inscription_height: Option<u64>,
+  pub(crate) first_inscription_height: Option<u32>,
   #[clap(long, help = "Limit index to <HEIGHT_LIMIT> blocks.")]
-  pub(crate) height_limit: Option<u64>,
+  pub(crate) height_limit: Option<u32>,
   #[clap(long, help = "Use index at <INDEX>.")]
   pub(crate) index: Option<PathBuf>,
   #[clap(long, help = "Track location of all satoshis.")]
@@ -79,7 +79,7 @@ impl Options {
     }
   }
 
-  pub(crate) fn first_inscription_height(&self) -> u64 {
+  pub(crate) fn first_inscription_height(&self) -> u32 {
     if self.chain() == Chain::Regtest {
       self.first_inscription_height.unwrap_or(0)
     } else if integration_test() {
